% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allfeat.R
\name{allfeat}
\alias{allfeat}
\title{all of the TAIR10 genes and sub-features with coordinates}
\format{one or more rows per locus/feature combination}
\description{
contains the start and stop coordinates in TAIR10 for each gene or subfeature. Index by "locus"
}
