% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allmeth.R
\name{allmeth}
\alias{allmeth}
\title{all of the TAIR10 genes and sub-features with coordinates and the number of methylated sites in the region}
\format{
one or more rows per locus/feature combination
}
\description{
contains the start and stop coordinates in TAIR10 for each gene or subfeature. Index by "locus".  The "methsites" column specifies how many methylated sites were found in the region given by the row coords
}
