% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneaccession.R
\name{geneaccession}
\alias{geneaccession}
\title{Just holds accession numbers and corresponding ATI locus ids}
\format{one row per SALK_Line}
\description{
three columns: 'idGeneAccession' (index), 'Accession_idAccession' (SALK Lines/Ecotypes),'Gene_idGene' (ATI ids)
}
