% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_impact_chr4.R
\name{high_impact_chr4}
\alias{high_impact_chr4}
\title{polymorph1001 variant data for chr4 loci}
\format{
one or more rows per locus/variant-type/strain(line) combination
}
\description{
Contains the variant information for all 1001 genome lines for each of the loci that were disrupted in the chr4 experiment
}
