% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/independent.R
\name{independent}
\alias{independent}
\title{Wide Format Version of additional quantitative information on genes}
\format{one row per SALK_Line}
\description{
mostly gene expression but also protein information and some ages of genes and insert locations
}

