% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methData.R
\name{methData}
\alias{methData}
\title{Locations and positions of methylation sites in the TAIR10 genome}
\format{
one row per site.
}
\description{
Has three columns that specify the chromosome, position, and type of methylation in the genome
}
