% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutationCat.R
\name{mutationCat}
\alias{mutationCat}
\title{takes a vector of accession numbers and returns a vector of mut. categories}
\usage{
mutationCat(accession, catnames = c("Columbia", "TDNA-mutant",
  "Phytometer"))
}
\arguments{
\item{accession}{vector of accession numbers}

\item{catnames}{vector of length 3 for the names of control, mutant, and phytometer, respectively}
}
\description{
takes a vector of accession numbers and returns a vector of mut. categories
}
