% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenowide.R
\name{phenowide}
\alias{phenowide}
\title{Wide Format Version of The unpak Phenotypes}
\format{one row per individual}
\description{
This data frame contains the typical type of dataset: one row per individual with lots of columns
}

