% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_chrom.R
\name{plotChrom}
\alias{plotChrom}
\title{plot a phenotype along a chromosome}
\usage{
plotChrom(
  pheno,
  adjust = c(0, 1, 2, 3)[2],
  chromo = 1,
  meta = c("3"),
  span = 1,
  lines = NULL
)
}
\arguments{
\item{pheno}{is a phenotype}

\item{adjust}{is an integer that describes how phenotypes are adjusted. 0=raw, 1=Columbia mean, 2=Phytometer mean, 3=All plant mean}

\item{chromo}{integer 1-5 corresponding to Arabidopsis chromosome number}

\item{meta}{the meta experiment default  experiment 3.  Can be a vector}

\item{span}{amount of line smoothing.  smaller is wigglyier}
}
\description{
plot a phenotype along a chromosome
}
