% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SalkPos.R
\name{SalkPos}
\alias{SalkPos}
\title{dataframe with the insert positions of all the Salk Lines}
\format{
one row per accession
}
\description{
Contains the insert position of salk TDNA insert lines relative to TAIR10
}
