% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slideWin.R
\name{slideWinMeth}
\alias{slideWinMeth}
\title{make a sliding window of methylation density}
\usage{
slideWinMeth(chr = 1, sz = 1e+05)
}
\arguments{
\item{chr}{chromosome number as a numeric value}

\item{sz}{size of sliding window}
}
\description{
make a sliding window of methylation density
}
