% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spread.R
\name{spread}
\alias{spread}
\title{takes a long phenotype object (with plantIDs) and spreads it into wide format}
\usage{
spread(dat)
}
\arguments{
\item{dat}{the phenolong frame or a subset of phenolong with all the same columns and headers}
}
\description{
takes a long phenotype object (with plantIDs) and spreads it into wide format
}
