% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdna.R
\name{tdna}
\alias{tdna}
\title{TDNRatio data indexed by Accession_idAccession}
\format{
one or more rows per Accession
}
\description{
contains the ratio data used to ID single inserts.  AreaRatio is the pre-calculated data for each line.
}
