% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thousandAndOne.R
\name{thousandAndOne}
\alias{thousandAndOne}
\title{1001 genome accession information}
\format{
one row per natural accession
}
\description{
Information downloaded from the 1001 genomes about the natural accession lines.  Notably this includes genetic cluster from the 2016 Cell paper and also lat and long for each collection
}
