% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variants_chr4.R
\name{variants}
\alias{variants}
\title{polymorph1001 variant data for loci whith phenotypes}
\format{
one or more rows per locus/variant-type/strain(line) combination
}
\description{
Contains the variant information for all 1001 genome lines for each of
the loci that have been phenotyped in unpak.

To use, you will need to use the function 'loadRDS()' like this:\\

variants = readRDS(paste0(system.file(package="unpakathon"),"extdata/variants.RDS"))
}
