% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneaccession.R
\name{geneaccession}
\alias{geneaccession}
\title{Just holds accession numbers and corresponding ATI locus ids}
\format{
one row per SALK_Line
}
\description{
three columns: 'idGeneAccession' (index), 'accession' (SALK Lines/Ecotypes),'locus' (ATI ids)
}
