% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenolong.R
\name{phenolong}
\alias{phenolong}
\title{Long Format Version of The unpak Phenotypes}
\format{one row per observation}
\description{
This data frame contains the following columns: "plantID" "accession" "meta.experiment" "facility" "Gene_idGene" "treatment" "value" "experiment" "phenotype".  The value column contains the actual observations the rest of the columns are used in classifying the observation
}

